refactor: move create_inbox fixture to inbox_fixture.py and update references
