+++ invenio_notify_test/test_tasks.py
@@ -6,12 +6,13 @@ from invenio_notify import constants
 from invenio_notify.records.models import NotifyInboxModel, EndorsementMetadataModel
 from invenio_notify.tasks import inbox_processing, mark_as_processed
 from invenio_notify_test.conftest import create_notification_data
+from invenio_notify_test.inbox_fixture import create_inbox
 
 
 def test_mark_as_processed(db, superuser_identity, create_inbox):
     """Test the mark_as_processed function."""
     # Create a test inbox record
-    inbox = create_inbox(record_id='r1')
+    inbox = create_inbox(recid='r1')
 
     # Initially, process_date should be None
     assert inbox.process_date is None
@@ -70,7 +71,7 @@ def test_inbox_processing_record_not_found(db, superuser_identity, create_inbox)
 
     # Create inbox record with notification pointing to non-existent record
     inbox = create_inbox(
-        record_id=recid,
+        recid=recid,
         raw=json.dumps(notification_data)
     )
